fix checkstyle and undeprecate/move tests
